+++ pgvector/sqlalchemy/__init__.py
@@ -1,4 +1,6 @@
 from .halfvec import Halfvec
+from .sparsevec import Sparsevec
 from .vector import Vector
+from ..utils import SparseVec
 
-__all__ = ['Vector', 'Halfvec']
+__all__ = ['Vector', 'Halfvec', 'Sparsevec']