+++ tests/test_sqlalchemy.py
@@ -1,5 +1,5 @@
 import numpy as np
-from pgvector.sqlalchemy import Vector, Halfvec
+from pgvector.sqlalchemy import Vector, Halfvec, Sparsevec, SparseVec
 import pytest
 from sqlalchemy import create_engine, insert, inspect, select, text, MetaData, Table, Column, Index, Integer
 from sqlalchemy.exc import StatementError
@@ -21,6 +21,7 @@ class Item(Base):
     id = mapped_column(Integer, primary_key=True)
     embedding = mapped_column(Vector(3))
     half_embedding = mapped_column(Halfvec(3))
+    sparse_embedding = mapped_column(Sparsevec(3))
 
 
 Base.metadata.drop_all(engine)
@@ -44,7 +45,7 @@ def create_items():
     ]
     session = Session(engine)
     for i, v in enumerate(vectors):
-        session.add(Item(id=i + 1, embedding=v, half_embedding=v))
+        session.add(Item(id=i + 1, embedding=v, half_embedding=v, sparse_embedding=SparseVec.from_dense(v)))
     session.commit()